Rest api integrated for creating quote

+++ quotes/api/views.py
@@ -1,10 +1,14 @@
 from rest_framework import status
 from rest_framework.response import Response
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
+from rest_framework.permissions import IsAuthenticated
+
+
 from rest_framework.pagination import PageNumberPagination
-from rest_framework.generics import ListAPIView
+from rest_framework.generics import ListAPIView, CreateAPIView
 from quotes.models import Quote
-from quotes.api.serializers import QuotesSerializer, QuoteDetailSerializer
+from registration.models import User
+from quotes.api.serializers import QuotesSerializer, QuoteDetailSerializer, QuoteCreateSerializer
 from django.shortcuts import get_object_or_404
 
 # @api_view(['GET',])
@@ -29,9 +33,13 @@ def apiQuoteDetailView(request, quote_id):
         serializer=QuoteDetailSerializer(quote)
         return Response(serializer.data)
 
-@api_view(['PUT',])
+@api_view(['GET','PUT'])
+@permission_classes((IsAuthenticated,))
 def apiQuoteEditView(request, quote_id):
     quote=get_object_or_404(Quote,id=quote_id)
+    user=request.user
+    if quote.author!=user:
+        return Response({'response':"you don't have permission to edit the quote!"})
     if request.method== "PUT":
         serializer=QuoteDetailSerializer(quote, data=request.data)
         data={}
@@ -40,10 +48,18 @@ def apiQuoteEditView(request, quote_id):
             data['success']="update successful"
             return Response(data=data)
         return Response(serializer.errors, status=status.HTTP_404_BAD_REQUEST)
+    else:
+        serializer=QuoteDetailSerializer(quote)
+        return Response(serializer.data)
+
 
 @api_view(['DELETE',])
+@permission_classes((IsAuthenticated,))
 def apiQuoteDeleteView(request, quote_id):
     quote=get_object_or_404(Quote,id=quote_id)
+    user=request.user
+    if quote.author!=user:
+        return Response({'response':"you don't have permission to delete the quote!"})
     if request.method== "DELETE":
         operation=quote.delete()
         data={}
@@ -52,3 +68,26 @@ def apiQuoteDeleteView(request, quote_id):
         else:
             data['failure']="delete failed"
         return Response(data=data)
+
+
+# @api_view(['GET','POST',])
+# @permission_classes((IsAuthenticated,))
+# def apiQuoteCreateView(request):
+#     author=request.user
+#     quote=Quote(author=author)
+#     if request.method== "POST":
+#         serializer=QuoteCreateSerializer(quote, data=request.data)
+#         if serializer.is_valid():
+#             serializer.save()
+#             return Response(serializer.data, status=status.HTTP_201_CREATED)
+#         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+#     else:
+#         serializer=QuoteCreateSerializer()
+#         return Response(serializer.data)
+
+
+
+class apiQuoteCreateView(CreateAPIView):
+    serializer_class = QuoteCreateSerializer
+    def perform_create(self, serializer):
+        serializer.save(author=self.request.user)
